tt/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.0.0
pytesseract==0.3.10
networkx==3.1
matplotlib==3.7.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
djangorestframework-simplejwt==5.3.0
django-filter==23.3
python-magic==0.4.27
pydantic==2.5.2
openai==1.3.0 